requirements.txt
python-dotenv
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
xgboost
pymongo
certifi
dill
pyyaml
-e .